[REFACTOR] Error, Success  Enum 경로 변경
