+++ CarrotServer/src/main/java/CarrotServer/controller/ClubController.java
@@ -5,8 +5,8 @@ import CarrotServer.common.response.ApiResponse;
 import CarrotServer.controller.request.ProfileRequestDTO;
 import CarrotServer.controller.response.ClubDetailResponseDTO;
 import CarrotServer.controller.response.ClubResponseDTO;
-import CarrotServer.exception.Error;
-import CarrotServer.exception.Success;
+import CarrotServer.exception.enums.Error;
+import CarrotServer.exception.enums.Success;
 import CarrotServer.service.ClubService;
 import CarrotServer.service.ProfileService;
 import io.swagger.v3.oas.annotations.Operation;
