+++ CarrotServer/src/main/java/CarrotServer/controller/LifeController.java
@@ -3,8 +3,8 @@ package CarrotServer.controller;
 import CarrotServer.common.response.ApiResponse;
 import CarrotServer.controller.response.LifeResponseDTO;
 import CarrotServer.common.enums.LifeCategory;
-import CarrotServer.exception.Error;
-import CarrotServer.exception.Success;
+import CarrotServer.exception.enums.Error;
+import CarrotServer.exception.enums.Success;
 import CarrotServer.service.LifeService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
